project setup: fix GO_TEST_PARALLEL to be less noisy the first time
With the new behavior to auto-configure the git submodules, the
`Makefile` prints out something that looks like an error when it's run
the first time, because the `GO_TEST_PARALLEL` calculation is trying to
divide a `''` by `2`. This sets a sensible default for the dividend so
that the line won't print out what looks like an error.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>