cloudsql instances: prefix all instance names with Kind
See #585 for more context. Previously, cloudsql instance names defaulted
to being a generated UUID. However, on the GCP side, cloudsql instance
names must begin with a letter, so any UUIDs starting with a number
would fail instance creation. This change prefixes the instance names
with the Kind so that they won't be affected by the naming restriction
anymore.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>