Make writeConnectionSecretToRef optional
https://github.com/crossplaneio/crossplane/issues/719

The resource claim reconciler (and API definitions) consider
writeConnectionSecretToRef to be optional, but the managed resource reconciler
fails if it is not specified. This change aligns the codebase on the reference
being optional. Managed resources that do not specify a secret reference will
provision successfully without publishing their connection details to a secret.

Signed-off-by: Nic Cope <negz@rk0n.org>