Return to runing reference resolution before observation
It turns out certain managed resources use part of their spec to observe their
external resource, and that part of the spec can depend on a reference. GCP
connections are an example of this; there's no "get connections" call so instead
we list all connections within a network, which is a spec field that could be
set by a reference.

Signed-off-by: Nic Cope <negz@rk0n.org>