Move password generation into its own package
https://github.com/sethvargo/go-password

I considered using the above library, but its API is a little overkill for our
needs. I chose not to keep the previous GeneratePassword implementation because
I felt the length argument was a unintuitive. The argument specified the length
of the randomly generated bytes, which were then base64 encoded resulting in a
~33% overhead. So for example GeneratePassword(20) generated a 27 character
password.

Signed-off-by: Nic Cope <negz@rk0n.org>