Add a simpler cross resource reference API, with support for reference selection
This commit deprecates (but still supports) the contemporary cross resource
reference API, in which API types must satisfy the resource.AttributeReferencer
type. It introduces a new implementation that supports both references (which
explicitly name a managed resource to derive a field value from) and selectors
(which set a referencer based on simple rules).

Signed-off-by: Nic Cope <negz@rk0n.org>