Update late initialized resources
A design goal of the managed.Reconciler implementation was to insulate folks who
are authoring an ExternalClient implementation from needing to interact with the
Kubernetes API directly. The separation of concerns is intended to be:

* ExternalClient accepts a managed resource struct and reconciles it with the
  provider. It may update the struct it is passed, but otherwise does not
  interact with the Kubernetes API.
* ExternalConnector interacts with the Kubernetes API in order to produce an
  ExternalClient.
* Reconciler orchestrates the ExternalConnector and ExternalClient, propagating
  updates from the ExeternalClient back to the Kubernetes API as necessary.

The desire to late initialize managed resources sullied this separation of
concern (or single responsibility) by plumbing a Kubernetes client down into the
ExternalClient implementations. I believe this was done because:

* We must update the resource spec before we update its status, or changes made
  to the status will be overwritten.
* We don't want to update the resource spec on every reconcile if it's possible
  to avoid doing so.

This small change to the managed resource reconciler allows an ExternalClient to
signal that it has late initialized a managed resource, and thus defer updating
that managed resource to the Reconciler. I believe this change is backward
compatible (ExternalClient implementations could continue to 'manually' late
initialize the resource and not return ResourceLateInitialized).

Signed-off-by: Nic Cope <negz@rk0n.org>