Remove pkg/test/integration
* https://github.com/crossplane/crossplane-runtime/pull/89
* https://github.com/search?q=%22github.com%2Fcrossplane%2Fcrossplane-runtime%2Fpkg%2Ftest%2Fintegration%22&type=code

This was added in the above PR, but we never really made use of it.
Based on the above GitHub search, I think every consumer of the package
(except https://github.com/vshn/crossplane-service-broker?) are ancient,
stale forks of kubevela, oam-kubernetes-runtime, or provider-gcp. Modern
versions of these packages no longer use this one.

If we wanted to be really conservative we could mark this deprecated,
but I lean toward just removing it.

Signed-off-by: Nic Cope <nicc@rk0n.org>